+++ quickstart-rag/complex-query/AzureSearchRagComplexQuery.csproj
